# api-gateway/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Gateway",
    "description": "Documentação da API Gateway"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/login": {
      "post": {
        "summary": "Realiza login do usuário admin",
        "tags": ["autenticacao"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": { "type": "string" },
                "password": { "type": "string" }
              },
              "required": ["username", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login realizado com sucesso"
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "summary": "Realiza logout do usuário autenticado",
        "tags": ["autenticacao"],
        "responses": {
          "200": {
            "description": "Logout realizado com sucesso"
          },
          "500": {
            "description": "Erro ao realizar logout"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Retorna a lista de usuários",
        "tags": ["usuarios"],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um novo usuário",
        "tags": ["usuarios"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Retorna um usuário específico",
        "tags": ["usuarios"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      },
      "put": {
        "summary": "Atualiza um usuário",
        "tags": ["usuarios"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do usuário"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      },
      "patch": {
        "summary": "Atualiza parcialmente um usuário",
        "tags": ["usuarios"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do usuário"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário parcialmente atualizado com sucesso",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      },
      "delete": {
        "summary": "Deleta um usuário",
        "tags": ["usuarios"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "204": {
            "description": "Usuário deletado com sucesso"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      }
    },
    "/auth": {
      "post": {
        "summary": "Autentica um usuário na API SOAP",
        "tags": ["autenticacao"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": { "type": "string" },
                "password": { "type": "string" }
              },
              "required": ["username", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Autenticação realizada com sucesso"
          }
        }
      }
    }
  },
"definitions": {
  "User": {
    "type": "object",
    "properties": {
      "id": { 
        "type": "integer",
        "readOnly": true
      },
      "name": { "type": "string" },
      "email": { "type": "string" },
      "password": { "type": "string" }
    },
    "required": ["name", "email", "password"]
  }
}

}